refactor(#5): split deal documents section into row, list, and dialogs
Extract format helpers, kind select, dropzone, upload/rename/delete dialogs, and document row/list so the cockpit section stays a thin composer.
