+++ app/dashboard/deals/cockpit/deal-document-delete-dialog.tsx
@@ -1,0 +1,57 @@
+'use client'
+
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/components/ui/alert-dialog'
+import { COPY } from '@/lib/copy'
+
+export function DealDocumentDeleteDialog({
+  open,
+  fileName,
+  pending,
+  onClose,
+  onConfirm,
+}: {
+  open: boolean
+  fileName: string
+  pending: boolean
+  onClose: () => void
+  onConfirm: () => void
+}) {
+  return (
+    <AlertDialog
+      open={open}
+      onOpenChange={(next) => {
+        if (!next) onClose()
+      }}
+    >
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>{COPY.deals.cockpit.documentsDeleteConfirm}</AlertDialogTitle>
+          <AlertDialogDescription>
+            {fileName} wird aus dem Deal und dem Speicher entfernt.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel disabled={pending}>Abbrechen</AlertDialogCancel>
+          <AlertDialogAction
+            disabled={pending}
+            onClick={(e) => {
+              e.preventDefault()
+              onConfirm()
+            }}
+          >
+            {pending ? 'Löschen …' : 'Löschen'}
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  )
+}
